[client] | refactor(svgs): 분리된 SVG 파일을 한 파일로 병합
svgs에 각 폴더로 분리되어있는 SVG 컴포넌트를 하나의 파일로 수정하였습니다.
