--- packages/client/svgs/GitHubSVG/index.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const GitHubSVG = props => (
-  <svg width={24} height={23.413} {...props}>
-    <path
-      d="M22.39 5.977a11.948 11.948 0 00-4.367-4.367A11.737 11.737 0 0012 0a11.738 11.738 0 00-6.023 1.609A11.946 11.946 0 001.61 5.976 11.738 11.738 0 000 12a11.662 11.662 0 002.289 7.055A11.753 11.753 0 008.2 23.39a.7.7 0 00.625-.109.611.611 0 00.2-.469l-.008-.844q-.008-.797-.008-1.39l-.359.062a4.585 4.585 0 01-.867.055 6.607 6.607 0 01-1.083-.109 2.426 2.426 0 01-1.047-.468 1.983 1.983 0 01-.687-.961l-.156-.36a3.9 3.9 0 00-.492-.8 1.885 1.885 0 00-.68-.594l-.109-.078a1.147 1.147 0 01-.2-.188.857.857 0 01-.141-.219q-.047-.11.078-.18a1.013 1.013 0 01.453-.07l.312.047a2.281 2.281 0 01.773.375 2.522 2.522 0 01.758.812 2.754 2.754 0 00.867.977 1.84 1.84 0 001.023.336 4.443 4.443 0 00.891-.078 3.108 3.108 0 00.7-.235 2.532 2.532 0 01.766-1.61 10.7 10.7 0 01-1.6-.281 6.377 6.377 0 01-1.469-.61 4.206 4.206 0 01-1.258-1.047 5.031 5.031 0 01-.82-1.641 7.8 7.8 0 01-.32-2.344 4.557 4.557 0 011.236-3.214 4.2 4.2 0 01.109-3.187 2.189 2.189 0 011.344.211 9.4 9.4 0 011.3.6q.414.25.664.422a11.284 11.284 0 016 0l.594-.375a8.409 8.409 0 011.437-.688 2.048 2.048 0 011.266-.172 4.16 4.16 0 01.125 3.187 4.558 4.558 0 011.235 3.219 7.889 7.889 0 01-.32 2.351 4.835 4.835 0 01-.828 1.641 4.368 4.368 0 01-1.266 1.039 6.389 6.389 0 01-1.469.609 10.688 10.688 0 01-1.6.282 2.783 2.783 0 01.812 2.219v3.3a.622.622 0 00.2.469.679.679 0 00.617.109 11.751 11.751 0 005.914-4.336A11.664 11.664 0 0024 11.999a11.748 11.748 0 00-1.61-6.022z"
-      fill="#fff"
-    />
-  </svg>
-);
-
-export default GitHubSVG;
