--- packages/client/svgs/HotlistSVG/index.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const HotlistSVG = props => (
-  <svg width={24} height={24} {...props}>
-    <path
-      d="M13.5.67a21.493 21.493 0 01.74 4.8 3.47 3.47 0 01-3.41 3.73A3.64 3.64 0 017.2 5.47l.03-.36A13.768 13.768 0 004 14a8 8 0 0016 0A17.115 17.115 0 0013.5.67zM11.71 19a3.176 3.176 0 01-3.22-3.14 3.132 3.132 0 012.81-3.12 7.712 7.712 0 004.62-2.58 13.953 13.953 0 01.59 4.04 4.8 4.8 0 01-4.8 4.8z"
-      fill="#02cf5d"
-    />
-    <path d="M0 0h24v24H0z" fill="none" />
-  </svg>
-);
-
-export default HotlistSVG;
