--- packages/client/svgs/LatestSVG/index.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-const LatestSVG = props => (
-  <svg width={24} height={24} {...props}>
-    <g opacity={0.7}>
-      <path d="M0 0h24v24H0z" fill="none" />
-      <path
-        d="M16.24 7.76A5.974 5.974 0 0012 6v6l-4.24 4.24a6 6 0 108.48-8.48zM12 2a10 10 0 1010 10A10 10 0 0012 2zm0 18a8 8 0 118-8 8 8 0 01-8 8z"
-        fill="#fff"
-      />
-    </g>
-  </svg>
-);
-
-export default LatestSVG;
