--- packages/client/svgs/LogoSVG/index.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-const LogoSVG = props => (
-  <svg
-    width={props.width || 121.859}
-    height={props.height || 27.422}
-    viewBox="0 0 122 28"
-    {...props}
-  >
-    <path
-      data-name="wedev();"
-      d="M0 6.657h3.531l1.5 10.547L6.89 8.376h2.2l2.109 8.8q.438-2.859.6-5.344t.164-4.187v-.984h3.391q-.422 3.7-1.031 7.719t-1.26 7.214H9.891l-1.953-8.312-1.844 8.312H2.828zM24.391 22.11a7.78 7.78 0 01-3.841-.938 6.667 6.667 0 01-2.648-2.7 8.52 8.52 0 01-.961-4.148 9.356 9.356 0 01.906-4.234 6.652 6.652 0 012.516-2.789 6.878 6.878 0 013.641-.977 6.357 6.357 0 013.242.836 5.866 5.866 0 012.289 2.484 8.64 8.64 0 01.844 3.977q0 .375-.062 1.063l-.031.438h-9.8a5.013 5.013 0 001.281 3.086 3.425 3.425 0 002.5 1.039 4.984 4.984 0 002.078-.477 4.516 4.516 0 001.719-1.336l1.986 1.988a7.1 7.1 0 01-5.659 2.688zm2.406-9.609a4.094 4.094 0 00-.367-1.734 3.11 3.11 0 00-1.047-1.273 2.667 2.667 0 00-1.57-.477 2.809 2.809 0 00-2.172.945 4.357 4.357 0 00-1.031 2.539zm12.281 9.577a6.191 6.191 0 01-3.234-.867 5.957 5.957 0 01-2.3-2.609 9.67 9.67 0 01-.844-4.242 9.853 9.853 0 01.87-4.266 6.72 6.72 0 012.32-2.8 5.578 5.578 0 013.156-.969 4.62 4.62 0 011.844.422 5.013 5.013 0 011.609 1.078V-.003h3.906v.359a.867.867 0 00-.219.414 3.693 3.693 0 00-.031.758l-.061 18.266a6.511 6.511 0 00.063.961 2.687 2.687 0 00.281.839H42.95a2.112 2.112 0 01-.258-.766 10.238 10.238 0 01-.036-1.034 3.4 3.4 0 01-1.484 1.7 4.18 4.18 0 01-2.094.584zm.516-2.984q2.937 0 2.938-5.141a7.92 7.92 0 00-.625-3.437 2.256 2.256 0 00-2.219-1.281 2.745 2.745 0 00-2.523 1.406 7.089 7.089 0 00-.837 3.653 6.647 6.647 0 00.492 2.75 3.564 3.564 0 001.23 1.566 2.778 2.778 0 001.544.484zm16.797 3.016a7.78 7.78 0 01-3.841-.938 6.667 6.667 0 01-2.648-2.7 8.52 8.52 0 01-.961-4.148 9.356 9.356 0 01.906-4.234 6.652 6.652 0 012.516-2.789 6.878 6.878 0 013.641-.977 6.357 6.357 0 013.242.836 5.866 5.866 0 012.289 2.484 8.64 8.64 0 01.844 3.977q0 .375-.062 1.063l-.031.438h-9.8a5.013 5.013 0 001.281 3.086 3.425 3.425 0 002.5 1.039 4.984 4.984 0 002.078-.477 4.516 4.516 0 001.719-1.336l1.986 1.988a7.1 7.1 0 01-5.659 2.688zm2.406-9.609a4.094 4.094 0 00-.367-1.734 3.11 3.11 0 00-1.047-1.273 2.667 2.667 0 00-1.57-.477 2.809 2.809 0 00-2.172.945 4.357 4.357 0 00-1.031 2.539zm19.609-5.844a10.734 10.734 0 01-.492 2.617q-.414 1.3-1.211 3.336l-3.547 8.984h-2.937L64.55 6.657h3.563l3.731 10.953 2.141-5.422q.625-1.8.938-3.047a10.212 10.212 0 00.313-2.484zM92.45 27.422a14.947 14.947 0 01-4.93-3.312 15.33 15.33 0 01-3.266-4.867 14.337 14.337 0 01-1.148-5.617 14.119 14.119 0 011.109-5.508 14.772 14.772 0 013.18-4.734A14.883 14.883 0 0192.262.126l.828 2.781a12.132 12.132 0 00-4.93 4.438 11.887 11.887 0 00-1.727 6.375 12.488 12.488 0 00.817 4.468 11.825 11.825 0 002.359 3.828 11.663 11.663 0 003.672 2.688zm6.656-2.719a11.663 11.663 0 003.675-2.687 11.825 11.825 0 002.359-3.828 12.488 12.488 0 00.813-4.469 11.887 11.887 0 00-1.727-6.375 12.132 12.132 0 00-4.93-4.437l.828-2.781a14.883 14.883 0 014.867 3.258 14.772 14.772 0 013.18 4.734 14.119 14.119 0 011.109 5.508 14.337 14.337 0 01-1.148 5.617 15.33 15.33 0 01-3.266 4.867 14.947 14.947 0 01-4.93 3.313zm20.097-12.671a2.389 2.389 0 01-1.156-.3 2.367 2.367 0 01-.875-.8 1.966 1.966 0 01-.328-1.086 1.928 1.928 0 01.328-1.078 2.391 2.391 0 01.875-.8 2.389 2.389 0 011.156-.3 2.526 2.526 0 011.141.281 2.421 2.421 0 01.914.781 1.885 1.885 0 01.352 1.109 1.966 1.966 0 01-.328 1.086 2.351 2.351 0 01-.883.8 2.494 2.494 0 01-1.196.307zm-2.937 13.343a9.994 9.994 0 001.508-1.742 2.87 2.87 0 00.6-1.4.47.47 0 00-.133-.3 5.82 5.82 0 00-.414-.422 5.863 5.863 0 01-.719-.781 1.309 1.309 0 01-.266-.8 2.245 2.245 0 01.3-1.141 2.127 2.127 0 01.836-.812 2.5 2.5 0 011.227-.3 2.4 2.4 0 011.3.383 2.869 2.869 0 01.984 1.055 3.056 3.056 0 01.375 1.516 5.662 5.662 0 01-.937 2.836 20.129 20.129 0 01-2.828 3.523z"
-      fill="#fff"
-    />
-  </svg>
-);
-
-export default LogoSVG;
