+++ packages/client/svgs/index.tsx
@@ -1,0 +1,112 @@
+import React from 'react';
+
+export const ArrowDropDown = props => (
+  <svg width={24} height={24} {...props}>
+    <path d="M7 10l5 5 5-5z" />
+    <path d="M0 0h24v24H0z" fill="none" />
+  </svg>
+);
+
+export const CloudSVG = props => (
+  <svg width={24} height={24} {...props}>
+    <path d="M0 0h24v24H0z" fill="none" />
+    <path
+      d="M19.35 10.04a7.492 7.492 0 00-14-2A6 6 0 006 20h13a4.986 4.986 0 00.35-9.96zM14 13v4h-4v-4H7l5-5 5 5z"
+      fill="#fff"
+    />
+  </svg>
+);
+
+export const FavoriteSVG = props => {
+  const { fill } = props;
+
+  return fill ? (
+    <svg width={24} height={24} viewBox="0 0 24 24" {...props}>
+      <path d="M0 0h24v24H0z" fill="none" />
+      <path d="M12 21.35l-1.45-1.32C5.4 15.36 2 12.28 2 8.5 2 5.42 4.42 3 7.5 3c1.74 0 3.41.81 4.5 2.09C13.09 3.81 14.76 3 16.5 3 19.58 3 22 5.42 22 8.5c0 3.78-3.4 6.86-8.55 11.54L12 21.35z" />
+    </svg>
+  ) : (
+    <svg width={24} height={24} viewBox="0 0 24 24" {...props}>
+      <path d="M0 0h24v24H0z" fill="none" />
+      <path d="M16.5 3c-1.74 0-3.41.81-4.5 2.09C10.91 3.81 9.24 3 7.5 3 4.42 3 2 5.42 2 8.5c0 3.78 3.4 6.86 8.55 11.54L12 21.35l1.45-1.32C18.6 15.36 22 12.28 22 8.5 22 5.42 19.58 3 16.5 3zm-4.4 15.55l-.1.1-.1-.1C7.14 14.24 4 11.39 4 8.5 4 6.5 5.5 5 7.5 5c1.54 0 3.04.99 3.57 2.36h1.87C13.46 5.99 14.96 5 16.5 5c2 0 3.5 1.5 3.5 3.5 0 2.89-3.14 5.74-7.9 10.05z" />
+    </svg>
+  );
+};
+
+export const GitHubSVG = props => (
+  <svg width={24} height={23.413} {...props}>
+    <path
+      d="M22.39 5.977a11.948 11.948 0 00-4.367-4.367A11.737 11.737 0 0012 0a11.738 11.738 0 00-6.023 1.609A11.946 11.946 0 001.61 5.976 11.738 11.738 0 000 12a11.662 11.662 0 002.289 7.055A11.753 11.753 0 008.2 23.39a.7.7 0 00.625-.109.611.611 0 00.2-.469l-.008-.844q-.008-.797-.008-1.39l-.359.062a4.585 4.585 0 01-.867.055 6.607 6.607 0 01-1.083-.109 2.426 2.426 0 01-1.047-.468 1.983 1.983 0 01-.687-.961l-.156-.36a3.9 3.9 0 00-.492-.8 1.885 1.885 0 00-.68-.594l-.109-.078a1.147 1.147 0 01-.2-.188.857.857 0 01-.141-.219q-.047-.11.078-.18a1.013 1.013 0 01.453-.07l.312.047a2.281 2.281 0 01.773.375 2.522 2.522 0 01.758.812 2.754 2.754 0 00.867.977 1.84 1.84 0 001.023.336 4.443 4.443 0 00.891-.078 3.108 3.108 0 00.7-.235 2.532 2.532 0 01.766-1.61 10.7 10.7 0 01-1.6-.281 6.377 6.377 0 01-1.469-.61 4.206 4.206 0 01-1.258-1.047 5.031 5.031 0 01-.82-1.641 7.8 7.8 0 01-.32-2.344 4.557 4.557 0 011.236-3.214 4.2 4.2 0 01.109-3.187 2.189 2.189 0 011.344.211 9.4 9.4 0 011.3.6q.414.25.664.422a11.284 11.284 0 016 0l.594-.375a8.409 8.409 0 011.437-.688 2.048 2.048 0 011.266-.172 4.16 4.16 0 01.125 3.187 4.558 4.558 0 011.235 3.219 7.889 7.889 0 01-.32 2.351 4.835 4.835 0 01-.828 1.641 4.368 4.368 0 01-1.266 1.039 6.389 6.389 0 01-1.469.609 10.688 10.688 0 01-1.6.282 2.783 2.783 0 01.812 2.219v3.3a.622.622 0 00.2.469.679.679 0 00.617.109 11.751 11.751 0 005.914-4.336A11.664 11.664 0 0024 11.999a11.748 11.748 0 00-1.61-6.022z"
+      fill="#fff"
+    />
+  </svg>
+);
+
+export const HotlistSVG = props => (
+  <svg width={24} height={24} {...props}>
+    <path
+      d="M13.5.67a21.493 21.493 0 01.74 4.8 3.47 3.47 0 01-3.41 3.73A3.64 3.64 0 017.2 5.47l.03-.36A13.768 13.768 0 004 14a8 8 0 0016 0A17.115 17.115 0 0013.5.67zM11.71 19a3.176 3.176 0 01-3.22-3.14 3.132 3.132 0 012.81-3.12 7.712 7.712 0 004.62-2.58 13.953 13.953 0 01.59 4.04 4.8 4.8 0 01-4.8 4.8z"
+      fill="#02cf5d"
+    />
+    <path d="M0 0h24v24H0z" fill="none" />
+  </svg>
+);
+
+export const LatestSVG = props => (
+  <svg width={24} height={24} {...props}>
+    <g opacity={0.7}>
+      <path d="M0 0h24v24H0z" fill="none" />
+      <path
+        d="M16.24 7.76A5.974 5.974 0 0012 6v6l-4.24 4.24a6 6 0 108.48-8.48zM12 2a10 10 0 1010 10A10 10 0 0012 2zm0 18a8 8 0 118-8 8 8 0 01-8 8z"
+        fill="#fff"
+      />
+    </g>
+  </svg>
+);
+
+export const LogoSVG = props => (
+  <svg
+    width={props.width || 121.859}
+    height={props.height || 27.422}
+    viewBox="0 0 122 28"
+    {...props}
+  >
+    <path
+      data-name="wedev();"
+      d="M0 6.657h3.531l1.5 10.547L6.89 8.376h2.2l2.109 8.8q.438-2.859.6-5.344t.164-4.187v-.984h3.391q-.422 3.7-1.031 7.719t-1.26 7.214H9.891l-1.953-8.312-1.844 8.312H2.828zM24.391 22.11a7.78 7.78 0 01-3.841-.938 6.667 6.667 0 01-2.648-2.7 8.52 8.52 0 01-.961-4.148 9.356 9.356 0 01.906-4.234 6.652 6.652 0 012.516-2.789 6.878 6.878 0 013.641-.977 6.357 6.357 0 013.242.836 5.866 5.866 0 012.289 2.484 8.64 8.64 0 01.844 3.977q0 .375-.062 1.063l-.031.438h-9.8a5.013 5.013 0 001.281 3.086 3.425 3.425 0 002.5 1.039 4.984 4.984 0 002.078-.477 4.516 4.516 0 001.719-1.336l1.986 1.988a7.1 7.1 0 01-5.659 2.688zm2.406-9.609a4.094 4.094 0 00-.367-1.734 3.11 3.11 0 00-1.047-1.273 2.667 2.667 0 00-1.57-.477 2.809 2.809 0 00-2.172.945 4.357 4.357 0 00-1.031 2.539zm12.281 9.577a6.191 6.191 0 01-3.234-.867 5.957 5.957 0 01-2.3-2.609 9.67 9.67 0 01-.844-4.242 9.853 9.853 0 01.87-4.266 6.72 6.72 0 012.32-2.8 5.578 5.578 0 013.156-.969 4.62 4.62 0 011.844.422 5.013 5.013 0 011.609 1.078V-.003h3.906v.359a.867.867 0 00-.219.414 3.693 3.693 0 00-.031.758l-.061 18.266a6.511 6.511 0 00.063.961 2.687 2.687 0 00.281.839H42.95a2.112 2.112 0 01-.258-.766 10.238 10.238 0 01-.036-1.034 3.4 3.4 0 01-1.484 1.7 4.18 4.18 0 01-2.094.584zm.516-2.984q2.937 0 2.938-5.141a7.92 7.92 0 00-.625-3.437 2.256 2.256 0 00-2.219-1.281 2.745 2.745 0 00-2.523 1.406 7.089 7.089 0 00-.837 3.653 6.647 6.647 0 00.492 2.75 3.564 3.564 0 001.23 1.566 2.778 2.778 0 001.544.484zm16.797 3.016a7.78 7.78 0 01-3.841-.938 6.667 6.667 0 01-2.648-2.7 8.52 8.52 0 01-.961-4.148 9.356 9.356 0 01.906-4.234 6.652 6.652 0 012.516-2.789 6.878 6.878 0 013.641-.977 6.357 6.357 0 013.242.836 5.866 5.866 0 012.289 2.484 8.64 8.64 0 01.844 3.977q0 .375-.062 1.063l-.031.438h-9.8a5.013 5.013 0 001.281 3.086 3.425 3.425 0 002.5 1.039 4.984 4.984 0 002.078-.477 4.516 4.516 0 001.719-1.336l1.986 1.988a7.1 7.1 0 01-5.659 2.688zm2.406-9.609a4.094 4.094 0 00-.367-1.734 3.11 3.11 0 00-1.047-1.273 2.667 2.667 0 00-1.57-.477 2.809 2.809 0 00-2.172.945 4.357 4.357 0 00-1.031 2.539zm19.609-5.844a10.734 10.734 0 01-.492 2.617q-.414 1.3-1.211 3.336l-3.547 8.984h-2.937L64.55 6.657h3.563l3.731 10.953 2.141-5.422q.625-1.8.938-3.047a10.212 10.212 0 00.313-2.484zM92.45 27.422a14.947 14.947 0 01-4.93-3.312 15.33 15.33 0 01-3.266-4.867 14.337 14.337 0 01-1.148-5.617 14.119 14.119 0 011.109-5.508 14.772 14.772 0 013.18-4.734A14.883 14.883 0 0192.262.126l.828 2.781a12.132 12.132 0 00-4.93 4.438 11.887 11.887 0 00-1.727 6.375 12.488 12.488 0 00.817 4.468 11.825 11.825 0 002.359 3.828 11.663 11.663 0 003.672 2.688zm6.656-2.719a11.663 11.663 0 003.675-2.687 11.825 11.825 0 002.359-3.828 12.488 12.488 0 00.813-4.469 11.887 11.887 0 00-1.727-6.375 12.132 12.132 0 00-4.93-4.437l.828-2.781a14.883 14.883 0 014.867 3.258 14.772 14.772 0 013.18 4.734 14.119 14.119 0 011.109 5.508 14.337 14.337 0 01-1.148 5.617 15.33 15.33 0 01-3.266 4.867 14.947 14.947 0 01-4.93 3.313zm20.097-12.671a2.389 2.389 0 01-1.156-.3 2.367 2.367 0 01-.875-.8 1.966 1.966 0 01-.328-1.086 1.928 1.928 0 01.328-1.078 2.391 2.391 0 01.875-.8 2.389 2.389 0 011.156-.3 2.526 2.526 0 011.141.281 2.421 2.421 0 01.914.781 1.885 1.885 0 01.352 1.109 1.966 1.966 0 01-.328 1.086 2.351 2.351 0 01-.883.8 2.494 2.494 0 01-1.196.307zm-2.937 13.343a9.994 9.994 0 001.508-1.742 2.87 2.87 0 00.6-1.4.47.47 0 00-.133-.3 5.82 5.82 0 00-.414-.422 5.863 5.863 0 01-.719-.781 1.309 1.309 0 01-.266-.8 2.245 2.245 0 01.3-1.141 2.127 2.127 0 01.836-.812 2.5 2.5 0 011.227-.3 2.4 2.4 0 011.3.383 2.869 2.869 0 01.984 1.055 3.056 3.056 0 01.375 1.516 5.662 5.662 0 01-.937 2.836 20.129 20.129 0 01-2.828 3.523z"
+      fill="#fff"
+    />
+  </svg>
+);
+
+export const ProfileSVG = props => (
+  <svg width={24} height={24} {...props}>
+    <path
+      d="M12 12a4 4 0 10-4-4 4 4 0 004 4zm0 2c-2.67 0-8 1.34-8 4v2h16v-2c0-2.66-5.33-4-8-4z"
+      fill="#fff"
+    />
+    <path d="M0 0h24v24H0z" fill="none" />
+  </svg>
+);
+
+export const SearchSVG = props => (
+  <svg width={23} height={24} {...props}>
+    <path
+      d="M14.785 14h-.745l-.264-.27a6.706 6.706 0 001.48-4.23A6.322 6.322 0 009.128 3 6.322 6.322 0 003 9.5 6.322 6.322 0 009.128 16a5.9 5.9 0 003.988-1.57l.255.28v.79l4.714 4.99L19.49 19zm-5.657 0a4.371 4.371 0 01-4.242-4.5A4.371 4.371 0 019.128 5a4.371 4.371 0 014.243 4.5A4.371 4.371 0 019.128 14z"
+      fill="#fff"
+    />
+    <path d="M0 0h23v24H0z" fill="none" />
+  </svg>
+);
+
+export const TopicsSVG = props => (
+  <svg width={24} height={24} {...props}>
+    <g opacity={0.7}>
+      <path d="M0 0h24v24H0z" fill="none" />
+      <path
+        d="M17.63 5.84A1.994 1.994 0 0016 5L5 5.01A2 2 0 003 7v10a2 2 0 002 1.99L16 19a1.994 1.994 0 001.63-.84L22 12l-4.37-6.16z"
+        fill="#fff"
+      />
+    </g>
+  </svg>
+);
