+++ packages/client/views/Main/index.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import Grid from '@material-ui/core/Grid';
 
 import * as S from './styles';
-import HotlistSVG from '../../svgs/HotlistSVG/';
 import Layout from '../../components/Layout';
 import Video from '../../components/Video';
+import { HotlistSVG } from '../../svgs';
 
 const Main: React.FunctionComponent = () => {
   return (
